fix namespacing of generators
